--- a/.github/workflows/test.yml
+++ b/.github/workflows/test.yml
@@ -6,7 +6,7 @@ jobs:
   unit_tests:
     strategy:
       matrix:
-        go-version: [1.18.x, 1.17.x, 1.12.x]
+        go-version: [1.18.x, 1.17.x, 1.16.x]
         os: [ubuntu-latest]
     runs-on: ${{ matrix.os }}
     steps: