--- a/config.go
+++ b/config.go
@@ -11,23 +11,14 @@ import (
 
 // Config contains VRF parameters.
 type Config struct {
+	// the elliptic curve.
+	Curve elliptic.Curve
 	// a single nonzero octet specifying the ECVRF ciphersuite.
 	SuiteString byte
 	// number of points on curve divided by group order.
 	Cofactor byte
 	// create cryptographic hash function.
 	NewHasher func() hash.Hash
-	// function to calculate y^2.
-	Y2 func(c elliptic.Curve, x *big.Int) *big.Int
-	// function to calculate square root.
-	Sqrt func(c elliptic.Curve, s *big.Int) *big.Int
-}
-
-// DefaultSqrt is the default sqrt method. nil is returned if s is not a square.
-func DefaultSqrt(c elliptic.Curve, s *big.Int) *big.Int {
-	var r big.Int
-	if nil == r.ModSqrt(s, c.Params().P) {
-		return nil // s is not a square
-	}
-	return &r
+	// decompress the compressed public key into x and y coordinate.
+	Decompress func(c elliptic.Curve, pk []byte) (x, y *big.Int)
 }