--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,5 @@
 module github.com/vechain/go-ecvrf
 
 go 1.12
+
+require github.com/decred/dcrd/dcrec/secp256k1/v4 v4.0.1