```
benchstat master.txt pr.txt 
name                                old time/op    new time/op    delta
VRF/secp256k1sha256tai-proving-8       483µs ±19%     281µs ± 9%  -41.97%  (p=0.008 n=5+5)
VRF/secp256k1sha256tai-verifying-8     645µs ± 7%     411µs ± 2%  -36.28%  (p=0.008 n=5+5)
VRF/p256sha256tai-proving-8            257µs ±11%     192µs ± 0%  -25.36%  (p=0.008 n=5+5)
VRF/p256sha256tai-verifying-8          353µs ± 8%     302µs ± 0%  -14.26%  (p=0.008 n=5+5)

name                                old alloc/op   new alloc/op   delta
VRF/secp256k1sha256tai-proving-8      15.5kB ±49%     4.9kB ± 0%  -68.42%  (p=0.008 n=5+5)
VRF/secp256k1sha256tai-verifying-8    15.8kB ±31%     5.0kB ± 0%  -68.24%  (p=0.016 n=5+4)
VRF/p256sha256tai-proving-8           12.0kB ±22%     9.1kB ± 0%  -24.10%  (p=0.008 n=5+5)
VRF/p256sha256tai-verifying-8         18.7kB ±12%    19.5kB ± 0%     ~     (p=0.127 n=5+5)

name                                old allocs/op  new allocs/op  delta
VRF/secp256k1sha256tai-proving-8        512 ±100%        95 ± 0%  -81.45%  (p=0.008 n=5+5)
VRF/secp256k1sha256tai-verifying-8       526 ±62%       114 ± 0%  -78.34%  (p=0.008 n=5+5)
VRF/p256sha256tai-proving-8              383 ±45%       294 ± 0%     ~     (p=0.127 n=5+5)
VRF/p256sha256tai-verifying-8            498 ±30%       515 ± 0%     ~     (p=0.683 n=5+5)
```